backend/requirements.txt
fastapi
sqlalchemy
uvicorn
python-multipart
python-dotenv
together
